refactor: extract FallbackPolicyTests nested doubles to TestDoubles (#306)
Moves TestQuery (+ Result), TestQueryHandlerWithCatchAllFallback,
TestQueryHandlerWithFormatExceptionFallback, and
TestQueryHandlerWithoutFormatExceptionFallback verbatim into
TestDoubles/ (one type per file). FallbackPolicyTests now references
them via a using; the Moq factory/registry usage is unchanged (swapped
in Task 8). No nested double class declarations remain in the file.

Satisfies FR3 (extraction half), FR5, AC6, AC7. Structural; suite green
at baseline.
