+++ test/Paramore.Darker.Core.Tests/TestDoubles/TestQuery.cs
@@ -1,0 +1,7 @@
+namespace Paramore.Darker.Core.Tests.TestDoubles
+{
+    internal class TestQuery : IQuery<TestQuery.Result>
+    {
+        internal class Result { }
+    }
+}
